cicero-app: alteração do tipo de dado de long para short em InfectedCreateDTO

+++ cicero-app/src/main/java/com/duende/cicero_app/dto/infected/InfectedCreateDTO.java
@@ -9,9 +9,9 @@ public record InfectedCreateDTO(
         String name,
         Integer determinationTotal,
         AttributeHealth attributeHealth,
-        Long generationId,
+        Short generationId,
         String generationDescription,
-        Long majorEventId,
-        Long occupationId
+        Short majorEventId,
+        Short occupationId
 ) {
 }
